dinner works in own class
