+++ lib/index.js
@@ -1,9 +1,7 @@
 "use strict"
-
 var Breakfast = require('./breakfast.js')
-var Lunch = require('./lunch')
-
-var allFoods = document.getElementById('myTable')
+var Lunch = require('./lunch.js')
+var Dinner = require('./dinner.js')
 
 function submitFood(food, calories, id){
   var newRow = document.createElement('tr');
@@ -23,9 +21,8 @@ function submitFood(food, calories, id){
   newRow.appendChild(deleteCell);
 
   $(newRow).data("id",id)
-
+  var allFoods = document.getElementById('myTable')
   allFoods.appendChild(newRow);
-
 };
 
 function updateTable() {
@@ -76,7 +73,6 @@ function updateExercise() {
 }
 
 updateExercise();
-
 
 // START OF BREAKFAST
 $('#submit-breakfast').on('click', function(event){
@@ -139,7 +135,6 @@ function submitBreakfast(food, calories, id){
   var breakfastTable = document.getElementById('breakfastTable')
 
   breakfastTable.appendChild(newRow);
-
 }
 
 function submitBreakfastCalories(total=0) {
@@ -181,10 +176,8 @@ function submitBreakfastCalories(total=0) {
   remainingRow.appendChild(emptyRemainingTd);
 
   breakfastTable.appendChild(remainingRow);
-
 }
 
-// sends local storage to submit breakfast that changes dom
 function updateBreakfast(){
   var breakfastCalorieTotal = 0
   if (localStorage.getItem('breakfast') === null) {
@@ -198,7 +191,6 @@ function updateBreakfast(){
     submitBreakfastCalories(breakfastCalorieTotal)
   }
 }
-
 // lunch starts here
 $('#submit-lunch').on('click', function(event){
   event.preventDefault();
@@ -240,7 +232,6 @@ function submitLunch(food, calories, id){
   var lunchTable = document.getElementById('lunchTable')
 
   lunchTable.appendChild(newRow);
-
 }
 
 function submitLunchCalories(total=0) {
@@ -263,9 +254,7 @@ function submitLunchCalories(total=0) {
 
   var lunchTable = document.getElementById('lunchTable')
   lunchTable.appendChild(newRow);
-
   // remaining row
-
   var remainingRow = document.createElement('tr');
   remainingRow.id = "lunch-calorie-total"
 
@@ -284,7 +273,6 @@ function submitLunchCalories(total=0) {
   remainingRow.appendChild(emptyRemainingTd);
 
   lunchTable.appendChild(remainingRow);
-
 }
 
 function updateLunch() {
@@ -314,24 +302,25 @@ function lunchChecker(checkBox){
   }
   clearChecked(checkBox);
 }
-
-
 // dinner starts here
-
-function deleteDinner(event, that) {
+$('#submit-dinner').on('click', function(event){
   event.preventDefault();
-  var dinner = JSON.parse(localStorage["dinner"]);
-  for (var i = 0; i < dinner.length; i++) {
-    if ($(that).parent().parent().data('id') === i){
-      that.parentElement.parentElement.remove();
-      dinner.splice(i,1)
-      dinnerJSON = JSON.stringify(dinner);
-      localStorage.setItem('dinner', dinnerJSON);
-    }
-   }
-   $("#dinnerTable > tr").remove();
-   updateDinner();
- };
+  var checkBox = $('.food-check')
+  dinnerChecker(checkBox);
+  $('#totals > tr').remove()
+  updateTotalTable()
+})
+
+$('#dinnerTable').on("click", ".fa-trash-o", function(event){
+  var that = this;
+  var dinnerClass = new Dinner()
+  dinnerClass.deleteDinner(event, that);
+  $("#dinnerTable > tr").remove();
+  updateDinner();
+  $('#totals > tr').remove()
+  updateTotalTable()
+});
+
 
 function submitDinner(food, calories, id){
   var newRow = document.createElement('tr');
@@ -357,7 +346,6 @@ function submitDinner(food, calories, id){
   dinnerTable.appendChild(newRow);
 }
 
-
 function submitDinnerCalories(total=0) {
   var newRow = document.createElement('tr');
   newRow.id = "dinner-calorie-total-tr"
@@ -378,9 +366,7 @@ function submitDinnerCalories(total=0) {
 
   var dinnerTable = document.getElementById('dinnerTable')
   dinnerTable.appendChild(newRow);
-
   // remaining row
-
   var remainingRow = document.createElement('tr');
   remainingRow.id = "dinner-calorie-total"
 
@@ -399,9 +385,7 @@ function submitDinnerCalories(total=0) {
   remainingRow.appendChild(emptyRemainingTd);
 
   dinnerTable.appendChild(remainingRow);
-
 }
-
 
 function updateDinner() {
   var dinnerCaloriesToatl = 0
@@ -417,35 +401,19 @@ function updateDinner() {
   }
 }
 
-function storeDinner(food, calories){
-  var dinnerJSON = localStorage.getItem('dinner')
-
-  if(dinnerJSON === null){
-    dinnerJSON = '[]';
-  };
-
-  var currentDinner = JSON.parse(dinnerJSON);
-  currentDinner.unshift({food: food, calories: calories});
-  dinnerJSON = JSON.stringify(currentDinner);
-  localStorage.setItem('dinner', dinnerJSON)
-
-  $("#dinnerTable > tr").remove();
-  updateDinner();
-}
-
 function dinnerChecker(checkBox){
   for (var i = 0; i < checkBox.length; i++) {
     if (checkBox[i].checked === true){
       var food = checkBox[i].parentElement.parentElement.parentElement.children[0].innerText;
       var calories = checkBox[i].parentElement.parentElement.parentElement.children[1].innerText;
-      storeDinner(food, calories)
+      var dinnerClass = new Dinner(food, calories)
+      dinnerClass.storeDinner()
+      $("#dinnerTable > tr").remove();
+      updateDinner();
     }
   }
   clearChecked(checkBox);
 }
-
-updateDinner();
-
 // snacks starts here
 
 function deleteSnacks(event, that) {
@@ -796,22 +764,8 @@ $(document).ready(function(){
   updateBreakfast();
   // lunch doc ready starts here
   updateLunch();
-
   // dinner doc ready starts here
-  $('#submit-dinner').on('click', function(event){
-    event.preventDefault();
-    var checkBox = $('.food-check')
-    dinnerChecker(checkBox);
-    $('#totals > tr').remove()
-    updateTotalTable()
-  })
-
-  $('#dinnerTable').on("click", ".fa-trash-o", function(event){
-    var that = this;
-    deleteDinner(event, that);
-    $('#totals > tr').remove()
-    updateTotalTable()
-  });
+  updateDinner();
   // snacks doc ready starts here
   $('#submit-snacks').on('click', function(event){
     event.preventDefault();
